backend/pyproject.toml
[project]
name = "cover-builder-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.2"
dependencies = [
    "alembic>=1.17.2",
    "fastapi>=0.127.0",
    "openai>=2.14.0",
    "psycopg[binary]>=3.3.2",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.45",
    "streamlit>=1.52.2",
    "uvicorn[standard]>=0.40.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "ruff>=0.14.10",
]
